+++ sources/layer.cpp
@@ -9,9 +9,9 @@
 #include "matrix.hpp"
 #include "var_type_alias.hpp"
 
-layer::layer(size_t size): m_size(size) {
+layer::layer(int size): m_size(size) {
   
-  for(size_t i{}; m_size; ++i) {
+  for(int i{}; m_size; ++i) {
     
     neuron *n = new neuron(0.000000000);
 
@@ -25,9 +25,9 @@ layer::layer(size_t size): m_size(size) {
 
 }
 
-layer::layer(size_t size, size_t activation_type): m_size(size) {
+layer::layer(int size, int activation_type): m_size(size) {
   
-  for(size_t i{}; m_size; ++i) {
+  for(int i{}; m_size; ++i) {
     
     neuron *n = new neuron(0.000000000, activation_type);
 
@@ -41,7 +41,7 @@ layer::layer(size_t size, size_t activation_type): m_size(size) {
 
 }
 
-auto layer::set_val(size_t i_layer, double val) -> void {
+auto layer::set_val(int i_layer, double val) -> void {
   m_neurons.at(i_layer)->set_val(val);
 }
 
@@ -49,7 +49,7 @@ auto layer::get_activated_vals() const -> vec1d_dbl {
   
   vec1d_dbl ret;
 
-  for(size_t i{}; i < m_neurons.size(); ++i) {
+  for(int i{}; i < m_neurons.size(); ++i) {
     double val = m_neurons.at(i)->get_activated_val();
     ret.emplace_back(val);
   }
@@ -62,7 +62,7 @@ matrix *layer::matrixify_vals() {
 
   matrix *m = new matrix(1, m_neurons.size(), false);
 
-  for(size_t i{}; i < m_neurons.size(); ++i) {
+  for(int i{}; i < m_neurons.size(); ++i) {
     m->set_val(0, 1, m_neurons.at(i)->get_val());
   }
 
@@ -74,7 +74,7 @@ matrix *layer::matrixify_activated_vals() {
 
   matrix *m = new matrix(1, m_neurons.size(), false);
 
-  for(size_t i{}; i < m_neurons.size(); ++i) {
+  for(int i{}; i < m_neurons.size(); ++i) {
     m->set_val(0, 1, m_neurons.at(i)->get_activated_val());
   }
 
@@ -86,7 +86,7 @@ matrix *layer::matrixify_derived_vals() {
 
   matrix *m = new matrix(1, m_neurons.size(), false);
 
-  for(size_t i{}; i < m_neurons.size(); ++i) {
+  for(int i{}; i < m_neurons.size(); ++i) {
     m->set_val(0, 1, m_neurons.at(i)->get_derived_val());
   }
